pyproject.toml
[project]
name = "port-cli"
version = "1.2.0"
description = "CLI for Port - import/export, migrate, and manage Port resources with a pluggable module architecture"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Port Labs", email = "support@getport.io" }
]
maintainers = [
    { name = "Port Labs", email = "support@getport.io" }
]
keywords = [
    "port",
    "cli",
    "devops",
    "developer-portal",
    "internal-developer-platform",
    "idp",
    "platform-engineering",
    "backstage-alternative"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
    "Typing :: Typed",
]

dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.9.0",
    "ruff>=0.3.0",
    "types-pyyaml>=6.0.12",
]

[project.scripts]
port = "port_cli.cli:run"

[project.urls]
Homepage = "https://github.com/port-labs/port-cli"
Documentation = "https://docs.getport.io"
Repository = "https://github.com/port-labs/port-cli"
Issues = "https://github.com/port-labs/port-cli/issues"
Changelog = "https://github.com/port-labs/port-cli/releases"
"Source Code" = "https://github.com/port-labs/port-cli"
"Bug Tracker" = "https://github.com/port-labs/port-cli/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/port_cli"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=port_cli --cov-report=term-missing --cov-report=html --cov-report=xml --cov-fail-under=58"
markers = [
    "unit: Unit tests for individual functions/classes",
    "integration: Integration tests for end-to-end workflows",
    "slow: Tests that take longer than 1 second",
]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

